+++ src/main/java/me/namuhuchutong/ladder/LadderGameApplication.java
@@ -1,5 +1,10 @@
 package me.namuhuchutong.ladder;
 
+import me.namuhuchutong.ladder.controller.LadderController;
+import me.namuhuchutong.ladder.domain.factory.ScaffoldRandomFactory;
+import me.namuhuchutong.ladder.ui.InputView;
+import me.namuhuchutong.ladder.ui.OutputView;
+
 public class LadderGameApplication {
 
     public static void main(String[] args){
