--- src/main/java/me/namuhuchutong/ladder/ScaffoldFactory.java
@@ -1,7 +1,0 @@
-package me.namuhuchutong.ladder;
-
-@FunctionalInterface
-public interface ScaffoldFactory {
-
-    LadderExpression createScaffold();
-}
