+++ src/main/java/me/namuhuchutong/ladder/domain/Names.java
@@ -1,4 +1,6 @@
-package me.namuhuchutong.ladder;
+package me.namuhuchutong.ladder.domain;
+
+import me.namuhuchutong.ladder.domain.wrapper.Name;
 
 import java.util.Arrays;
 import java.util.List;
