+++ src/main/java/me/namuhuchutong/ladder/domain/factory/ScaffoldFactory.java
@@ -1,0 +1,9 @@
+package me.namuhuchutong.ladder.domain.factory;
+
+import me.namuhuchutong.ladder.domain.wrapper.LadderExpression;
+
+@FunctionalInterface
+public interface ScaffoldFactory {
+
+    LadderExpression createScaffold();
+}
