+++ src/main/java/me/namuhuchutong/ladder/domain/wrapper/LadderExpression.java
@@ -1,4 +1,4 @@
-package me.namuhuchutong.ladder;
+package me.namuhuchutong.ladder.domain.wrapper;
 
 public abstract class LadderExpression {
