+++ src/test/java/me/namuhuchutong/ladder/HyphenTest.java
@@ -1,5 +1,7 @@
 package me.namuhuchutong.ladder;
 
+import me.namuhuchutong.ladder.domain.wrapper.Hyphen;
+import me.namuhuchutong.ladder.domain.wrapper.LadderExpression;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
