+++ src/test/java/me/namuhuchutong/ladder/RowTest.java
@@ -1,5 +1,8 @@
 package me.namuhuchutong.ladder;
 
+import me.namuhuchutong.ladder.domain.Row;
+import me.namuhuchutong.ladder.domain.factory.ScaffoldFactory;
+import me.namuhuchutong.ladder.domain.wrapper.Hyphen;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
