+++ src/test/java/me/namuhuchutong/ladder/RowsTest.java
@@ -1,5 +1,8 @@
 package me.namuhuchutong.ladder;
 
+import me.namuhuchutong.ladder.domain.Rows;
+import me.namuhuchutong.ladder.domain.factory.ScaffoldFactory;
+import me.namuhuchutong.ladder.domain.wrapper.EmptySpace;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
